requirements.txt
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flake8==7.0.0
idna==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
vk-api==11.9.9
